rename of CSS styles file
