+++ calminder/src/components/faq/FaqCategory.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from "./fqa.module.css";
+import styles from "./faq.module.css";
 import { useEffect, useState } from 'react';
 import FaqCard from './FaqCard';
 import { getFaqs } from '../../services/firebase.service';
